refactor : Like 컨트롤러 및 서비스 클래스 주석 제거 및 Impl 패턴 제거
